codes/pytest_example/requirements.txt
pytest
fastapi
uvicorn
requests
pytest-result-log
allure-pytest
pyyaml
jsonschema
jsonpath-ng